FIX: BigDecimal support in jdk1.5 negative scale - fix for IntegerDecimal class.
git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/trunk@2262 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.engine/src/org/jetel/data/primitive/IntegerDecimal.java
@@ -650,9 +650,9 @@ public final class IntegerDecimal implements Decimal {
 
     private double convertToInnerForm(double d) {
         if(scale >= 0) {
-            return (long) d * TENPOWERS[scale];
+            return d * TENPOWERS[scale];
         } else {
-            return (long) d / TENPOWERS[-scale];
+            return d / TENPOWERS[-scale];
         }
     }
